Adds a hook to specify unmarshal options.
This makes it possible to configure the parser to error out if
an unknown tagged field has been found in the byte stream
that is being unmarshaled to a go object.

NOTE: This change requires go 1.10.  It seems that we should all
upgrade, see: https://pocketgophers.com/when-should-you-upgrade-go/

For example like this (also see ExampleUnknown in yaml_test.go):

```
func ExampleUnknown() {
	type WithTaggedField struct {
		Field string `json:"field"`
	}
	y := []byte(`unknown: "hello"`)
	v := WithTaggedField{}
	fmt.Printf("%v\n", Unmarshal(y, &v, DisallowUnknownFields))
	// Ouptut:
	// unmarshaling JSON: while decoding JSON: json: unknown field "unknown"
}
```

If you want to manipulate the decoder used in the unmarshal function,
you can define a custom option to your liking and apply it in the
Unmarshal call.

The way I found out about this is kind of comical.  Some nonprintable
characters made their way into my YAML file.  You wouldn't see them in the
editor, but they'd be there.  Now when the YAML file is parsed, the
nonprintable chars together with the key characters were made into a
weird UTF-8 key (e.g. like "\u00c2some_key"), which would then
become a key unknown to the tags of my struct, and would silently get
dropped in the parsing steps.  So as result, you would get a perfectly
normal parse, except the data that you wanted to be in the struct are not
there.

This meant that silently dropping unknown keys is not always a good
idea, so I went to fix it, while retaining the same interface.  JSONOpt
type was fairly easy to add in the Unmarshal because it can be silently
ignored by the old code.