Add YAMLToJSONCustom to accept custom unmarshal func
We want to use YAMLToJSON with yaml.v3 and this is the most minimal
change that allows us to use yaml.v3 without breaking the interface of
the rest of the library.

Why v3? Because v3 changed how it treats boolean values:

> YAML 1.1 bools (yes/no, on/off) are supported as long as they are
> being decoded into a typed bool value. Otherwise they behave as a
> string. Booleans in YAML 1.2 are true/false only.
(https://github.com/go-yaml/yaml/blob/v3/README.md#compatibility)